+++ tinper-bee/bee-form-control/src/FormControl.js
@@ -44,8 +44,10 @@ class FormControl extends React.Component {
     }
 
     componentWillReceiveProps(nextProp) {
-        if (nextProp.value !== this.state.value) {
-            this.setState({value: nextProp.value});
+        if ("value" in nextProp) {
+            if (nextProp.value !== this.state.value) {
+                this.setState({value: nextProp.value});
+            }
         }
     }