+++ tinper-bee/bee-popover/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="3.0.7"></a>
+## [3.0.7](https://github.com/tinper-bee/bee-popover/compare/v3.0.6...v3.0.7) (2020-01-14)
+
+
+
 <a name="3.0.6"></a>
 ## [3.0.6](https://github.com/tinper-bee/bee-popover/compare/v3.0.5...v3.0.6) (2019-11-06)