+++ tinper-bee/bee-timepicker/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="2.0.12"></a>
+## [2.0.12](https://github.com/tinper-bee/bee-timepicker/compare/v2.0.11...v2.0.12) (2020-01-09)
+
+
+
 <a name="2.0.11"></a>
 ## [2.0.11](https://github.com/tinper-bee/bee-timepicker/compare/v2.0.10...v2.0.11) (2019-11-27)